requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.4
Werkzeug==2.3.7
opencv-python==4.6.0.66
picamera2==0.0.2
Flask-Migrate==4.0.0
Python-Dotenv==1.0.0
Email-Validator==1.2.1
OpenAI==0.0.0
